Composite Command - Money Transfer

+++ DesignPatters/Models/Proxy/IBankAccount.cs
@@ -2,7 +2,7 @@ namespace DesignPatters.Models.Proxy;
 
 public interface IBankAccount
 {
-    void Deposit(int amount);
-    bool Withdraw(int amount);
+    void Deposit(decimal amount);
+    bool Withdraw(decimal amount);
     string ToString();
 }
